+++ Blaster Master/Dome.h
@@ -4,6 +4,6 @@
 class Dome : public DynamicObject
 {
 public:
-	void Update(DWORD dt);
+	void Update(float dt);
 	void Render();
 };
